Fix containers faults on the update.
The is that both `currentPMMC` and `currentAppC` were pointing to the same address.
Because `for - range` placed all "containers" into the same memory address (compiler optimizations). So on each step of the loop, despite that the `c` contained different data, this data lied at the same memory address. ha ha