pitr: write binlog to s3 with io.Pipe
in previus implementation we was blocking on waiting for stderror.
This commit fixes this by providing buffer to stderror instead of
blocking on ioutill.ReadAll, which waits for io.EOF. But this error is
returned only at the end of command execution, which causing buffer grow
and pod eviction from k8s side.

Now, we are using io.Pipe and start writing to s3 immideatly to not
store any info inside pitr.
Plus, increase buffer size to make bigger reads, instead of 1kb chunks