pitr: check error buffer before closing pipe
We need this check because if any error occures, mysqlbinlog sends EOF
to file, and we process further as normal situation.

Now, before closing pipe, we check the error buffer, and if any error
happens, we cancel upload to s3 and return error correctly.

We will process the failed binlog on next collector iteration