Make peer-list handle SIGUSR1
The Docker image
percona/percona-xtradb-cluster-operator:$version-haproxy configures the
STOPSIGNAL to SIGUSR1 to shutdown HAProxy gracefully. peer-list needs to
handle this as well because it is bundled in the same container.
Otherwise peer-list will just ignore the signal which leads to the pod
staying in `Terminating` status for quite a while.

Closes K8SPXC-509