K8SPXC-985: Remove binlog filter (#1117)
We were trying to filter binlogs while applying to make storing binlogs
from multiple clusters in the same bucket possible. The filter is
implemented with the assumption that we can't have a different GTID from
the source cluster that we created the full backup from. This was wrong.
There are multiple ways that can lead GTID to change and making the
cluster that collects binlogs a replica of another is one of them.

We'll recommend users to use a separate storage/bucket/directory to
store the binlogs and remove the filter from PITR recoverer. All the
binlogs will be applied one by one.