K8SPXC-1030: Don't delete cert-manager certs by default (#1171)
* K8SPXC-1030: Don't delete cert-manager certs by default

We shouldn't set owner references to cert-manager objects if we don't
want to delete secrets too. This way, after the PXC cluster is deleted
issuers, certificates and their secrets will remain intact in the
cluster.

If users want to cleanup objects created for SSL, we introduce a new
finalizer: `delete-ssl`. If this finalizer is set, the operator will
delete secrets, certificates and issuers. Unfortunately, cert-manager
doesn't set owner reference to the secret it created and this behaviour
can only configured by command line flag in the controller. Since we
can't control how users deploy cert-manager to their clusters, we
shouldn't rely on this feature and cleanup certificates and secrets
altogether. Hopefully,
https://github.com/cert-manager/cert-manager/pull/5158 will merged and
we can configure this behaviour on certificate level.

* fix tests

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>