K8SPXC-1118: Monitor binlog gaps
Before these changes, binlog collector was printing seemingly harmless
log in case it detects a gap in binlogs. The log was too easy to miss
and users have little chance to understand that their latest backup
can't guarantee consistent recovery with PITR.

With these changes, operator will monitor the binlog gaps detected by
binlog collector. Collector will create a temporary file if a gap is
detected and the operator will check if the file exists by execing into
collector container. If file exists, operator will add a new condition
to backup that marks the backup "not suitable for consistent PITR".

If the user tries to restore (with PITR) to the backup which is marked as not
suitable, the operator will block the restore process and mark the
restore operation as failed. User needs to annotate the restore object
to force restore.