K8SPXC-1005: fix azure backup deletion
https://jira.percona.com/browse/K8SPXC-1005

There was a problem while deleting the `#innodb_redo` folder in azure backup.
The `azblob` package has problems with deleting blobs that contain special characters.
This issue is mentioned here: https://github.com/Azure/azure-sdk-for-go/issues/19475.

We should url escape the blob names during the delete request.

Also, in this PR, we keep finalizers in backup, if they were failed