K8SPXC-1546: Cache binlog->gtid set pairs
Currently, the binlog collector flushes binary logs every X seconds. X
is configurable by users through cr.yaml. Usually users want to run
collector every 60 seconds to collect any event happened in last minute
and meet their objectives for recovery. Running the collector every 60
seconds has the nasty side effect of creating a lot of binlogs. For
example, after running the collector for a month creates `30 * 24 * 60 =
43200` binlogs.

Collector connects to the PXC host that has the oldest binlog. That
means the host it collects binlogs change after binlogs are purged
and/or expired. For this reason, the collector mainly cares about the
GTID sets in each binlog file. In every run, collector first runs `SHOW
BINARY LOGS` to get the binlog list and runs `SELECT
get_gtid_set_by_binlog(?)` for each binlog in the list to assign a GTID
set to each binlog.

After running the collector for a month (and having 43200+ binlogs as a
result), collector lose the ability to recover from crash. Because it's
impossible to go through 43k+ binlogs in a 60 second time window. For
this reason, we decided to cache binlog->gtid set pairs.

With these changes, the collector will maintain a cache in json in the
same storage that binlogs are uploaded. When collector starts, it'll
first check the cache file and if it can't find the cache it'll ignore
all timeout until the cache is populated. Cache contains binlog names
and GTID sets for each PXC host.

There are some events that requires cache to be invalidated because they
affect the binlogs on hosts:
* After a restore, cache needs to be invalidated for all hosts. This is
  performed by the operator after restore succeeds. Operator simply
  deletes the cache file in binlog storage to make collector re-populate
  the cache.
* After a SST, cache needs to be invalidated for Joiner host. This is
  performed by the Joiner host itself. To achieve this, we use
  `wsrep_notify_cmd` to trigger our script whenever there's a status
  change. Triggered script sends a HTTP request to collector pod (using
  the Service we create for collector) with its hostname, and collector
  deletes cache entries for that host and update the cache file in
  storage.