K8SPXC-1376: Don't require matching passwords for restoring a backup (#2046)
* K8SPXC-1376: Don't require matching passwords for restoring a backup

This commit adds a post-restore step to change passwords in restored
backup so operator can continue reconciling cluster with existing
secrets and liveness/readiness work.

After backup is restored, operator will create a another job that starts
`mysqld` with `--skip-grant-tables` and change user passwords in the
database by using passwords in the secrets. If some system users are not
in the database, this job will create them. If users don't have some
grants, those will be added to users as well.

* fix prepare for encrypted clusters

* fix password escaping in test

* fix 5.7

* remove xtrace

* add unit test

* fix unit tests

* fix unit tests

* fix tests

---------

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>