K8SPXC-1347: PITR Restore Fails When spec.pxc.configuration.[mysqld] require_secure_transport=ON is set (#2063)
https://perconadev.atlassian.net/browse/K8SPXC-1347

Problem:
When the server is configured with require_secure_transport=ON, pitr
restore fails.

Cause:
When the server is configured with require_secure_transport=ON it
requires that all client connections use TLS. Operator uses internal
connections to connect to PXC node. The connection is configured with
tls: preferred which means "Use TLS if supported, fall back to plain if
not.". This was implemented in commit c2c600b7.
However, PITR uses the connection without tls: preferred, which means
it uses the default 'false' value (do not use TLS). This causes
connection failure.

Solution:
Configure connection used by PITR restore to use TLS if required by
the server.

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>