K8SPXC-735: Add ProxySQL scheduler support
This commit introduces changes to support pxc_scheduler_handler. The
code for scheduler can be found in github.com/percona/pxc_scheduler_handler.
Managing scheduler requires percona-scheduler-admin, a bash script
equivalent to proxy-admin. Its code can be found in github.com/percona/proxysql-admin-tool.

Scheduler is disabled by default to keep the old behavior. We might
consider enabling it by default in future releases. To enable scheduler,
it's enough to just set `proxysql.scheduler.enabled=true`:
```
proxysql:
  size: 3
  image: percona/percona-proxysql:2.7.3
  ...
  scheduler:
    enabled: true
```

Scheduler will set `pod-0` as the writer node and won't let users to
have more than 1 writer. `pod-0` will be the only writer as long is it's
available. If it's not available, ProxySQL will promote another pod to
be the writer.

There are a few options that users can tweak for their needs:
```
scheduler:
  enabled: true
  writerIsAlsoReader: true
  checkTimeoutMilliseconds: 2000
  successThreshold: 1
  failureThreshold: 3
  pingTimeoutMilliseconds: 1000
  nodeCheckIntervalMilliseconds: 2000
  maxConnections: 1000
```

* **writerIsAlsoReader**: If you want to exclude the writer from read
  set it to false. When the cluster will lose its last reader, the
  writer will be elected as Reader, no matter what.
* **checkTimeoutMilliseconds**: If checking a backend node (PXC) exceeds
  this timeout, it won't be processed.
* **successThreshold**: Number of retries the application should do before
  restoring a failed node.
* **failureThreshold**: Number of retries the application should do to
  put DOWN a failing node.
* **pingTimeoutMilliseconds**: The connection timeout (milliseconds)
  used to test the connection towards the PXC server.
* **nodeCheckIntervalMilliseconds**: How frequently the scheduler must run.
* **maxConnections**: Max number of connections from ProxySQL to the backend servers.

Full configuration options can be found in `build/proxysql_scheduler_config.tmpl`.

One important change in this implementation is disabling clustering in
ProxySQL. Since our tests showed that scheduler and ProxySQL cluster
don't play nicely together, we decided to disable ProxySQL clustering if
scheduler enabled. Setting up the cluster was done in `proxysql-monit`
sidecar container in ProxySQL pods. When scheduler is enabled, we remove
this container from the pod. All ProxySQL servers are responsible of
populating `mysql_servers` on their own which they do in `pxc-monit`
container.

Disabling cluster-mode in ProxySQL has a nasty side effect: If pod-0
becomes unavailable, one ProxySQL server can promote pod-1 and another
ProxySQL server can promote pod-2 during failover. This is okay from the
integrity standpoint thanks to PXC's conflict resolution, but it's not
okay from performance standpoint as conflict resolution will put
additional burden on PXC.

Because of this we need to ensure all ProxySQL servers will promote the
same pod in case of failover. We ensure this by setting different
weights for each individual pod:
```
hg_id	hostname	weight	status
10	some-name-pxc-0	600	ONLINE
10	some-name-pxc-1	999	ONLINE
10	some-name-pxc-2	998	ONLINE
11	some-name-pxc-0	1000000	ONLINE
8010	some-name-pxc-0	600	ONLINE
8010	some-name-pxc-1	999	ONLINE
8010	some-name-pxc-2	998	ONLINE
8011	some-name-pxc-0	1000000	ONLINE
8011	some-name-pxc-1	999	ONLINE
8011	some-name-pxc-2	998	ONLINE
```

`10` is the reader hostgroup, `11` is the writer hostgroup. `8010` is
used to configure reader options, in case there needs to be failover in
the reader hostgroup. `8011` is used to configure writer options, in
case there needs to be failover in the writer hostgroup.