K8SPXC-1511: Install keyring component for 8.4 (#1948)
* K8SPXC-1511: Install keyring component for 8.4

This commit adds data-at-rest-encryption support for PXC 8.4.

In 8.4, keyring needs to be installed using components rather than
plugins. For components to be installed, you need to generate manifest
file(s). Either you can use a global manifest file at
`/usr/sbin/mysqld.my` or a local one in `/var/lib/mysql`. To use the
local one, you still need to create a global manifest file containing
`{ "read_local_manifest": true }`. We have the global manifest file in
our PXC 8.4 images since February, 2025.

If `vaultSecretName` is defined in cr.yaml, operator automatically
configures PXC to use data-at-rest-encryption. In case of 8.4, operator:

1. Creates the local manifest file `/var/lib/mysql/mysqld.my` with
   contents `{ "components": "file://component_keyring_vault" }`
2. Copies vault secret file to `/var/lib/mysql/component_keyring_vault.cnf`.

> [!warning]
> Vault config needs to be named `component_keyring_vault.cnf` both for
PXC and PXB. They use the naming to detect keyring type.

If `vaultSecretName` is not defined in cr.yaml, PXC entrypoint will
clean up `mysqld.my` and/or `component_keyring_vault.cnf` if they exist.

Note that vault configuration type is changed in 8.4. For 5.7 and 8.0,
the config was an ini file. In 8.4, it's expected to be a json. The
template can be found at `e2e-tests/conf/vault-secret-84.yaml`

* fixes

---------

Co-authored-by: Viacheslav Sarzhan <slava.sarzhan@percona.com>